refactor(viewer): modify getCircleEditPoints location to Annotation
