+++ libs/insight-viewer/src/Annotation/utils/getCircleEditPoints.spec.ts
