+++ libs/insight-viewer/src/Annotation/utils/getCircleEditPoints.ts
